refactor: rename closing icon
