+++ src/components/icons/ClosingX.tsx
